mqtt_monitor/requirements.txt
# MQTT Monitoring System - Dependencies

# Core
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0

# MQTT Client
paho-mqtt>=1.6.1

# Web Server
eventlet>=0.33.0
gevent>=23.9.0

# Database (choose one or use in-memory)
# sqlite3 is built-in to Python

# Optional: Redis for advanced caching
# redis>=5.0.0

# Optional: Database ORM
# sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
